[llvm][cas] New mapped file design to fix filesystem assumptions
On the advice of filesystem experts, we are changing our approach to
the memory-mapped files used in the CAS. Specifically, we will no longer
attempt to resize the underlying file while holding an mmap, which does
not behave correctly on all filesystems (e.g. tmpfs)  and is formally
unspecified by POSIX. Ideally, we could simply update the mapped regions
whenever we resize, but in practice this is challenging to do portably,
cross-process, without locking the reading process' access, and without
requiring us to check every offset before reading it and trigger a
re-map if necessary (which is hard to test/maintain).

Instead, we will truncate the file to its maximum size immediately, and
resize it back to the allocated size when closing, if possible. On
filesystems with sparse file support, this does not require any
additional storage. See the comments in MappedFileRegionBumpPtr for
details.

Because the test suite is uniquely likley to (a) crash from time to
time, and (b) have many separate CAS instances for tests, we override
the usual index/data/actioncache file sizes to be at most 100 MB. This
should help mitigate the risks of having many large CAS instances left
around if testing on a filesystem without sparse file support.

rdar://105905301
(cherry picked from commit 0046c5d6fff17c3d0b5ab99cfc8c2920922c5474)

+++ llvm/test/CAS/Inputs/cas-creation-stress-test.sh
@@ -9,7 +9,10 @@ NUM_REPEAT=$4
 set -e
 
 for c in $(seq 1 $NUM_REPEAT); do
-  rm -rf $CAS_PATH
+  # Half the time, start from an existing CAS.
+  if [[ $(($c % 2)) -eq 0 ]]; then
+    rm -rf $CAS_PATH
+  fi
 
   pids=""
   for x in $(seq 1 10); do
@@ -20,4 +23,12 @@ for c in $(seq 1 $NUM_REPEAT); do
   for pid in $pids; do
     wait "$pid"
   done
+
+  cas_kb=$(du -ks $CAS_PATH | sed 's|\t.*$||')
+  limit_in_kb=$((100 * 1024))
+  # Check that the CAS is smaller than 100 MB, to ensure it was resized.
+  if [[ $cas_kb -gt $limit_in_kb ]]; then
+    echo "error: on-disk cas size ${cas_kb}kb was larger than ${limit_in_kb}kb" >&2
+    exit 1
+  fi
 done
